widen-system-uid task groups 2,6.1,6.2: Akka.Cluster + DData int->long system uid
- UniqueAddress(Address, long uid) + long Uid; GetHashCode via Uid.GetHashCode()
  (value-identical to the old int hash for legacy-range uids); CompareTo/Equals
  already long-safe
- ClusterMessages.proto UniqueAddress.uid uint32 -> uint64 (same varint wire
  type; binary-compatible for uids <= uint32.MaxValue; regenerated by Grpc.Tools)
- ClusterMessageSerializer: (uint)/(int) narrowing replaced with lossless
  (ulong)/(long) reinterpret casts at the proto boundary
- DData SerializationSupport.UniqueAddressFromProto: (int) narrowing removed
  (ReplicatorMessages.proto already int64)
- vclock node name verified string-only (ClusterDaemon.VclockName) - unchanged

Whole solution builds green.

+++ src/core/Akka.Cluster/Serialization/ClusterMessageSerializer.cs
@@ -578,14 +578,14 @@ namespace Akka.Cluster.Serialization
         {
             var message = new Proto.Msg.UniqueAddress();
             message.Address = AddressToProto(uniqueAddress.Address);
-            message.Uid = (uint)uniqueAddress.Uid;
+            message.Uid = (ulong)uniqueAddress.Uid;
             return message;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static UniqueAddress UniqueAddressFrom(Proto.Msg.UniqueAddress uniqueAddressProto)
         {
-            return new UniqueAddress(AddressFrom(uniqueAddressProto.Address), (int)uniqueAddressProto.Uid);
+            return new UniqueAddress(AddressFrom(uniqueAddressProto.Address), (long)uniqueAddressProto.Uid);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
